+++ src/api/devices/deviceTransformer.js
@@ -6,6 +6,7 @@ const transformDevice = device => ({
   type: device.deviceType,
   version: device.swVersion,
   status: device.status ? device.status : {},
+  id: device._id, // eslint-disable-line no-underscore-dangle
 });
 
 export { transformDevice }; // eslint-disable-line import/prefer-default-export