+++ src/api/gateways/gatewayTransformer.js
@@ -3,6 +3,7 @@
 const transformGateway = gateway => ({
   code: gateway.code,
   description: gateway.description,
+  id: gateway._id, // eslint-disable-line no-underscore-dangle
 });
 
 export { transformGateway }; // eslint-disable-line import/prefer-default-export