+++ src/data/mongodb/events/eventsProvider.js
@@ -1,37 +1,26 @@
 import { DataProvider } from '../data';
 
-const EventsProvider = ({ db, collectionName }) => ({
-  /**
-   * Returns all events belonging to the gateways passed as parameter
-   */
-  getEvents(gateways) {
-    return new Promise((resolve, reject) => {
-      db.collection(collectionName, (err, col) => {
-        if (err) {
-          reject(err);
-        }
-
-        col.find({
-          GatewayId: {
-            $in: gateways,
-          },
-        })
-          .toArray((error, docs) => {
-            if (error) {
-              reject(error);
-            } else {
-              resolve(docs);
-            }
-          });
-      });
-    });
-  },
-});
-
-export default function (db) {
+export default function (database) {
   const params = {
-    db,
+    db: database,
     collectionName: 'events',
   };
-  return Object.assign({}, DataProvider(params), EventsProvider(params));
+
+  const dataProvider = DataProvider(params);
+
+  const EventsProvider = () => ({
+    /**
+     * Returns all events belonging to the gateways passed as parameter
+     */
+    getEvents(gateways) {
+      return dataProvider.getMany({
+        GatewayId: {
+          $in: gateways,
+        },
+      });
+    },
+  });
+
+
+  return Object.assign({}, dataProvider, EventsProvider());
 }