+++ src/data/mongodb/index.js
@@ -1,6 +1,6 @@
 export { default as GatewaysProvider } from './gateways/gatewaysProvider';
 export { default as EventsProvider } from './events/eventsProvider';
 export { default as DevicesProvider } from './devices/devicesProvider';
-export { default as AlertsProdiver } from './alerts/alertsProvider';
+export { default as AlertsProvider } from './alerts/alertsProvider';
 export { default as DailyStatsProvider } from './stats/dailyStatsProvider';
 export { default as HourlyStatsProvider } from './stats/hourlyStatsProvider';